fix: Added mac support
Initial mac support by upgrading electron packager to the latest version, since the other version crashed with no custom icon, and changing the file locations for all binaries and files for mac to the userData directory. Some more testing needs to be done to get it to support mac after this commit.